# agent_api_http/openapi-full.yaml
openapi: 3.1.0
info:
  title: UniCore HTTP API
  description: ''
  license:
    name: Apache 2.0
  version: 0.0.0-semantically-released
servers:
- url: http://localhost:3033
  description: Local development
paths:
...
  /auth/par:
    post:
      tags:
      - OAuth 2.0
      - Protocol
      summary: Push an authorization request
      description: |-
        Handles the Pushed Authorization Request (PAR) endpoint as defined by
        [RFC 9126](https://www.rfc-editor.org/rfc/rfc9126.html), as well as the Interactive Authorization Request flow
        as defined in OpenID4VCI 1.1.

        Nested values, such as `authorization_details`, are JSON-encoded form values.
      operationId: auth_par
      requestBody:
        content:
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/AuthorizationRequestDto'
        required: true
      responses:
        '200':
          description: Interactive authorization response to an initial or follow-up interactive authorization request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InteractiveAuthorizationResponse'
        '201':
          description: Pushed authorization response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PushedAuthorizationResponse'
        '400':
          description: The request body is not `application/x-www-form-urlencoded`
        '422':
          description: The request body is not a valid authorization request
        '500':
          description: The authorization request could not be processed
...
  /credential_offer:
    get:
      tags:
      - OpenID4VCI
      - Protocol
      summary: Receive a credential offer
      description: |-
        Receives a credential offer from a credential issuer, as defined by
        [OpenID4VCI](https://openid.net/specs/openid-4-verifiable-credential-issuance-1_0.html#name-credential-offer-endpoint).
        `credential_offer` takes precedence when both parameters are present.
      operationId: credential_offer
      parameters:
      - name: credential_offer
        in: query
        description: JSON-encoded credential offer.
        required: false
        schema:
          type: string
      - name: credential_offer_uri
        in: query
        description: URL from which the credential offer can be retrieved.
        required: false
        schema:
          type: string
      responses:
        '200':
          description: Credential offer received
        '400':
          description: Neither parameter is present, or the credential offer is invalid
        '500':
          description: The credential offer could not be stored
...
components:
  schemas:
...
    AuthorizationDetailsClaim:
      type: object
      required:
      - path
      properties:
        mandatory:
          type: boolean
        path:
          $ref: '#/components/schemas/ClaimPathPointer'
    AuthorizationDetailsObject:
      type: object
      description: |-
        Represents an object of the `authorization_details` field of the `AuthorizationRequest` object in the Authorization Code Flow as
        and in the `TokenRequest` object. : https://openid.net/specs/openid-4-verifiable-credential-issuance-1_0.html#authorization-details
      required:
      - type
      - credential_configuration_id
      properties:
        claims:
          type:
          - array
          - 'null'
          items:
            $ref: '#/components/schemas/AuthorizationDetailsClaim'
        credential_configuration_id:
          type: string
        credential_identifiers:
          type:
          - array
          - 'null'
          items:
            type: string
        locations:
          type:
          - array
          - 'null'
          items:
            type: string
            format: uri
        type:
          $ref: '#/components/schemas/OpenidCredential'
    AuthorizationRequest:
      type: object
      description: |-
        The Authorization Request is used to request authorization as described here: https://openid.net/specs/openid-4-verifiable-credential-issuance-1_0-final.html#name-authorization-endpoint
        In combination with this important paragraph: https://openid.net/specs/openid-4-verifiable-credential-issuance-1_0-final.html#name-identifying-credentials-bei
        NOTE: this Authorization Request is not to be confused with the same-named request in OpenID for Verifiable Presentations.
      required:
      - response_type
      - client_id
      properties:
        authorization_details:
          type:
          - array
          - 'null'
          items:
            $ref: '#/components/schemas/AuthorizationDetailsObject'
        client_id:
          type: string
        code_challenge:
          type:
          - string
          - 'null'
        code_challenge_method:
          oneOf:
          - type: 'null'
          - $ref: '#/components/schemas/CodeChallengeMethod'
        issuer_state:
          type:
          - string
          - 'null'
        redirect_uri:
          type:
          - string
          - 'null'
          format: uri
        response_type:
          type: string
        scope:
          type:
          - string
          - 'null'
        state:
          type:
          - string
          - 'null'
    AuthorizationRequestDto:
      oneOf:
      - $ref: '#/components/schemas/InteractiveAuthorizationRequest'
      - $ref: '#/components/schemas/InteractiveAuthorizationFollowUpRequest'
      - $ref: '#/components/schemas/AuthorizationRequest'
...
    ClaimPathPointer:
      type: array
      items:
        oneOf:
        - type: string
          description: To address a particular claim within an object, append the key (claim name) to the array.
        - type: integer
          format: int64
          description: To address an element within an array, append the index to the array (as a non-negative, 0-based integer).
          minimum: 0
        - type: 'null'
          description: To address all elements within an array, append a null value to the array.
          default: null
        description: One segment of a Claim Path Pointer.
      minItems: 1
...
    CodeChallengeMethod:
      type: string
      description: PKCE code challenge methods accepted by the Authorization Server.
      enum:
      - S256
      - plain
...
    InteractionType:
      oneOf:
      - type: string
        description: The Wallet supports an OpenID4VP Presentation interaction.
        enum:
        - urn:openid:dcp:iae:openid4vp_presentation
      - type: string
        description: The Wallet supports a redirect to a web-based interaction.
        enum:
        - urn:openid:dcp:iae:redirect_to_web
      - type: string
        description: Custom interaction type defined by an extension.
      description: Interaction types supported by the Wallet as defined in Section 6.1.1.
    InteractiveAuthorizationFollowUpRequest:
      type: object
      description: |-
        A follow-up request to the Interactive Authorization Endpoint, as defined in Section 6.1.2.

        Follow-up requests include the `auth_session` value received most recently from the
        Authorization Server. Additional parameters depend on the interaction type.
      required:
      - auth_session
      properties:
        auth_session:
          type: string
          description: The auth_session value from the most recent IAE response.
        code_verifier:
          type:
          - string
          - 'null'
          description: |-
            The PKCE code verifier, required after a `urn:openid:dcp:iae:redirect_to_web`
            interaction if PKCE was used in the initial request.
        openid4vp_response:
          description: |-
            The OpenID4VP Authorization Response (JSON-encoded), present when responding
            to a `urn:openid:dcp:iae:openid4vp_presentation` interaction.
    InteractiveAuthorizationRequest:
      allOf:
      - $ref: '#/components/schemas/AuthorizationRequest'
      - type: object
        required:
        - interaction_types_supported
        properties:
          interaction_types_supported:
            type: string
            description: Comma-separated list of interaction types the Wallet supports.
      description: |-
        The initial request to the Interactive Authorization Endpoint, as defined in Section 6.1.1.

        Formed and sent in the same way as a PAR request (RFC 9126 Section 2.1), with the addition
        of the `interaction_types_supported` parameter.
    InteractiveAuthorizationResponse:
      type: object
      description: |-
        Response from the Interactive Authorization Endpoint, as defined in Section 6.2.

        The response indicates either that user interaction is required, that the authorization
        was completed successfully (with an authorization code), or an error.
      required:
      - status
      properties:
        auth_session:
          type:
          - string
          - 'null'
          description: |-
            A value that allows the Authorization Server to associate subsequent requests
            with the ongoing authorization request sequence. Must be included in follow-up requests.
        code:
          type:
          - string
          - 'null'
          description: The authorization code, present when `status` is `ok`.
        expires_in:
          type:
          - integer
          - 'null'
          format: int64
          description: The lifetime of the `request_uri` in seconds.
        openid4vp_request:
          description: |-
            An OpenID4VP Authorization Request for the Wallet to process.
            May contain either a plain request object or a signed request (`{"request": "eyJ..."}`).
        request_uri:
          type:
          - string
          - 'null'
          description: A request_uri for building an Authorization Request via browser redirect.
        status:
          $ref: '#/components/schemas/InteractiveAuthorizationStatus'
          description: Whether an additional interaction is required or the authorization has been completed.
        type:
          oneOf:
          - type: 'null'
          - $ref: '#/components/schemas/InteractionType'
            description: The interaction type required by the Authorization Server.
    InteractiveAuthorizationStatus:
      type: string
      description: The response status from the Interactive Authorization Endpoint.
      enum:
      - require_interaction
      - ok
...
    OpenidCredential:
      type: string
      description: Represents the `openid_credential` field of the `AuthorizationDetailsObject`.
      enum:
      - openid_credential
...
    PushedAuthorizationResponse:
      type: object
      description: Successful Pushed Authorization Request response as defined by RFC 9126.
      required:
      - request_uri
      - expires_in
      properties:
        expires_in:
          type: integer
          format: int64
        request_uri:
          type: string
...
  securitySchemes:
    access_token:
      type: http
      scheme: bearer
      description: Access token issued by the `/auth/token` endpoint.
...